# internal/api/spec/knot-openapi.yaml
openapi: 3.0.3
servers:
  - url: https://127.0.0.1
    description: Local server.
info:
  version: 1.0.0
  title: Knot API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: ../images/logo-light.svg
  description: |
    # Introduction
    The REST API is used for the CLI, Web interface and 3rd party integrations.
...
paths:
...
  /api/audit-logs:
    get:
      tags:
      - Audit Logs
      summary: Get Audit Logs
      description: Retrieve the audit logs.
      operationId: getAuditLogs
      parameters:
      - name: start
        in: query
        description: The item to start at.
        required: false
        schema:
          type: integer
          format: int32
      - name: max-items
        in: query
        description: The maximum number of items to return.
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        '200':
          description: A list of audited operations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogs'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/tunnels:
    get:
      tags:
        - Tunnels
      summary: Get Tunnels
      description: Retrieve all tunnels for the authenticated user.
      operationId: getTunnels
      responses:
        '200':
          description: A list of tunnels.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TunnelInfo'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/users:
    post:
      tags:
        - Users
      summary: Create User
      description: |
        Create a new user.

        Only admin users can create users.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/userRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    description: True if the user was created.
                  user_id:
                    type: string
                    format: uuid
                    description: The ID of the user.
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
      security: [BearerAuth: []]

    get:
      tags:
        - Users
      summary: Get Users
      description: Get the list of users within the system.
      operationId: listUsers
      parameters:
        - name: state
          in: query
          description: The sate of users to return
          required: false
          schema:
            type: string
            enum:
              - active
              - inactive
              - all
            default: all
        - name: zone
          in: query
          description: If given tracks the number of spaces the user has deployed in the given zone.
          required: false
          schema:
            type: string
        - name: local
          in: query
          description: If given and true then only users that have logged into a leaf node at least once will be returned.
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: The total number of users
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserInfo'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/groups:
    get:
      summary: Get Groups
      description: Retrieve all groups.
      operationId: getGroups
      tags:
        - Groups
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfoList'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    post:
      summary: Create a Group
      description: Create a new group.
      operationId: createGroup
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
      responses:
        '201':
          description: Group created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
      security: [BearerAuth: []]
...
  /api/roles:
    get:
      tags:
        - Roles
      summary: Get a List of Roles
      description: Get the list of roles defined within the system.
      operationId: listRoles
      responses:
        '200':
          description: A list of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id_role:
                      type: string
                      format: uuid
                    role_name:
                      type: string
        '401':
          $ref: '#/components/responses/unauthorized'
      security: [BearerAuth: []]

    post:
      summary: Create a role
      description: Create a new role.
      operationId: createRole
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '201':
          description: Group created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
      security: [BearerAuth: []]
...
  /api/sessions:
    get:
      tags:
        - Users
      summary: Get Sessions
      description: Retrieve all sessions for the authenticated user.
      operationId: GetSessions
      responses:
        '200':
          description: A list of sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/auth/web:
    post:
      tags:
        - Authorization
      summary: Login Web User
      description: |
        Authenticate a user with the server and create a web access token.

        A session is also opened and session cookie set within the response.
      operationId: WebLoginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: []
...
  /api/spaces:
    get:
      summary: Get Spaces
      description: Retrieve all spaces for a user.
      operationId: getSpaces
      tags:
        - Spaces
      parameters:
        - name: user_id
          in: query
          description: Optional user to limit the space list to, if not given then the current authenticated user will be used.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceInfoList'
        '401':
          $ref: '#/components/responses/unauthorized'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    post:
      summary: Create a Space
      description: Create a space for the authenticated user.
      operationId: createSpace
      tags:
        - Spaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceRequest'
      responses:
        '201':
          description: Space created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    description: True if the space was created.
                  space_id:
                    type: string
                    format: uuid
                    description: The ID of the created space.
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internal-server-error'
        '507':
          $ref: '#/components/responses/insufficient-storage'
      security: [BearerAuth: []]
  /api/spaces/{space_id}:
    delete:
      summary: Delete a Space
      description: Delete a space for the authenticated user.
      operationId: deleteSpace
      tags:
        - Spaces
      parameters:
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the space to delete.
      responses:
        '200':
          description: Successful operation
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/not-found'
        '423':
          $ref: '#/components/responses/locked'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    put:
      summary: Update a Space
      description: Update a space for the authenticated user.
      operationId: updateSpace
      tags:
        - Spaces
      parameters:
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the space to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceRequest'
      responses:
        '200':
          description: Space update successfully
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    get:
      summary: Get a Space
      description: Retrieve a specific space.
      operationId: getSpace
      tags:
        - Spaces
      parameters:
        - name: space_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the space to get.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceDefinition'
        '401':
          $ref: '#/components/responses/unauthorized'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/templates:
    get:
      summary: Get Templates
      description: Retrieve all templates.
      operationId: getTemplates
      tags:
        - Templates
      parameters:
        - name: user_id
          in: query
          description: If given, only templates available to that user will be returned.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateList'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    post:
      summary: Create a Template
      description: Create a new template.
      operationId: createTemplate
      tags:
        - Templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreateRequest'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    description: True if the template was created.
                  template_id:
                    type: string
                    format: uuid
                    description: The ID of the created template.
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
  /api/volumes:
    get:
      summary: Get Volumes
      description: Retrieve all volumes.
      operationId: getVolumes
      tags:
        - Volumes
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    description: The total number of volumes
                  volumes:
                    type: array
                    items:
                      $ref: '#/components/schemas/VolumeInfo'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    post:
      summary: Create a Volume
      description: Create a new volume.
      operationId: createVolume
      tags:
        - Volumes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolumeCreateRequest'
      responses:
        '201':
          description: Volume created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    description: True if the volume was created.
                  volume_id:
                    type: string
                    format: uuid
                    description: The ID of the created volume.
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GroupInfoList:
      type: object
      properties:
        count:
          type: integer
          example: 1
          description: The total number of groups.
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupInfo'
    GroupRequest:
      type: object
      properties:
        name:
          type: string
          example: "my group"
          maxLength: 64
          description: The name of the group.
        max_spaces:
          type: integer
          format: uint32
          description: The maximum number of spaces the group members can have.
        compute_units:
          type: integer
          format: uint32
          description: The number of compute units the group members get.
        storage_units:
          type: integer
          format: uint32
          description: The number of storage units the group members get.
        max_tunnels:
          type: integer
          format: uint32
          description: The maximum number of tunnels the group members can have.
...
    AuditLogs:
      type: object
      properties:
        count:
          type: integer
          description: The total number of audit logs.
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
...
    TunnelInfo:
      type: object
      properties:
        name:
          type: string
          description: The name of the tunnel.
        address:
          type: string
          description: The address of the tunnel.
...
  responses:
    bad-request:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    internal-server-error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...